+++ pymc3/distributions/__init__.py
@@ -80,6 +80,7 @@ from pymc3.distributions.distribution import (
 )
 from pymc3.distributions.mixture import Mixture, MixtureSameFamily, NormalMixture
 from pymc3.distributions.multivariate import (
+    CAR,
     Dirichlet,
     DirichletMultinomial,
     KroneckerNormal,
@@ -184,4 +185,5 @@ __all__ = [
     "Simulator",
     "fast_sample_posterior_predictive",
     "BART",
+    "CAR",
 ]